+++ android/src/main/java/com/theoplayer/source/SourceAdapter.kt
@@ -17,6 +17,8 @@ import com.theoplayer.android.api.source.addescription.GoogleImaAdDescription
 import com.theoplayer.android.api.player.track.texttrack.TextTrackKind
 import com.theoplayer.android.api.source.metadata.ChromecastMetadataImage
 import com.theoplayer.BuildConfig
+import com.theoplayer.android.api.ads.theoads.TheoAdDescription
+import com.theoplayer.android.api.ads.theoads.TheoAdsLayoutOverride
 import com.theoplayer.android.api.error.ErrorCode
 import com.theoplayer.android.api.source.AdIntegration
 import com.theoplayer.android.api.source.dash.DashPlaybackConfiguration
@@ -52,7 +54,15 @@ private const val PROP_ADS = "ads"
 private const val PROP_DASH = "dash"
 private const val PROP_DASH_IGNORE_AVAILABILITYWINDOW = "ignoreAvailabilityWindow"
 private const val PROP_HEADERS = "headers"
+private const val PROP_BACKDROP_DOUBLE_BOX = "backdropDoubleBox"
+private const val PROP_BACKDROP_LSHAPE = "backdropLShape"
+private const val PROP_CUSTOM_ASSET_KEY = "customAssetKey"
+private const val PROP_OVERRIDE_LAYOUT = "overrideLayout"
+private const val PROP_NETWORK_CODE = "networkCode"
+private const val PROP_USE_ID3 = "useId3"
+
 private const val ERROR_IMA_NOT_ENABLED = "Google IMA support not enabled."
+private const val ERROR_THEOADS_NOT_ENABLED = "THEOads support not enabled."
 private const val ERROR_UNSUPPORTED_CSAI_INTEGRATION = "Unsupported CSAI integration"
 private const val ERROR_MISSING_CSAI_INTEGRATION = "Missing CSAI integration"
 
@@ -252,6 +262,9 @@ class SourceAdapter {
         AdIntegration.GOOGLE_IMA.adIntegration -> parseImaAdFromJS(
           jsonAdDescription
         )
+        AdIntegration.THEO_ADS.adIntegration -> parseTheoAdFromJS(
+          jsonAdDescription
+        )
         else -> {
           throw THEOplayerException(
             ErrorCode.AD_ERROR,
@@ -283,6 +296,31 @@ class SourceAdapter {
     return GoogleImaAdDescription.Builder(source)
       .timeOffset(jsonAdDescription.optString(PROP_TIME_OFFSET))
       .build()
+  }
+
+  @Throws(JSONException::class)
+  private fun parseTheoAdFromJS(jsonAdDescription: JSONObject): TheoAdDescription {
+    if (!BuildConfig.EXTENSION_THEOADS) {
+      throw THEOplayerException(ErrorCode.AD_ERROR, ERROR_THEOADS_NOT_ENABLED)
+    }
+    return TheoAdDescription(
+      networkCode = jsonAdDescription.optString(PROP_NETWORK_CODE),
+      backdropDoubleBox = jsonAdDescription.optString(PROP_BACKDROP_DOUBLE_BOX),
+      backdropLShape = jsonAdDescription.optString(PROP_BACKDROP_LSHAPE),
+      customAssetKey = jsonAdDescription.optString(PROP_CUSTOM_ASSET_KEY),
+      overrideLayout = parseOverrideLayout(jsonAdDescription.optString(PROP_OVERRIDE_LAYOUT)),
+      useId3 = jsonAdDescription.optBoolean(PROP_USE_ID3),
+    )
+  }
+
+  private fun parseOverrideLayout(layout: String?): TheoAdsLayoutOverride? {
+    return when (layout) {
+      "single" -> TheoAdsLayoutOverride.SINGLE
+      "l-shape" -> TheoAdsLayoutOverride.LSHAPE
+      "double" -> TheoAdsLayoutOverride.DOUBLE
+      "single-if-mobile" -> null /* Not supported yet */
+      else -> null
+    }
   }
 
   @Throws(JSONException::class)