+++ example/src/screens/SampleExtendedScreen.tsx
@@ -1,0 +1,158 @@
+import * as React from 'react';
+import { useState } from 'react';
+import {
+  AirplayButton,
+  CastMessage,
+  CenteredControlBar,
+  CenteredDelayedActivityIndicator,
+  ChromecastButton,
+  ControlBar,
+  DEFAULT_THEOPLAYER_THEME,
+  FullscreenButton,
+  LanguageMenuButton,
+  MuteButton,
+  PipButton,
+  PlaybackRateSubMenu,
+  PlayButton,
+  QualitySubMenu,
+  SeekBar,
+  SettingsMenuButton,
+  SkipButton,
+  Spacer,
+  TimeLabel,
+  UiContainer,
+} from '@theoplayer/react-native-ui';
+import { PlayerConfiguration, PlayerEventType, THEOplayer, THEOplayerView } from 'react-native-theoplayer';
+
+import { Platform, SafeAreaView, StyleSheet } from 'react-native';
+import { SourceMenuButton, SOURCES } from '../custom/SourceMenuButton';
+import { BackgroundAudioSubMenu } from '../custom/BackgroundAudioSubMenu';
+import { PiPSubMenu } from '../custom/PipSubMenu';
+import { MediaCacheDownloadButton } from '../custom/MediaCacheDownloadButton';
+import { MediaCacheMenuButton } from '../custom/MediaCacheMenuButton';
+import { MediaCachingTaskListSubMenu } from '../custom/MediaCachingTaskListSubMenu';
+import { THEO_LICENSE } from '../sampleConfig';
+
+const LOG_TAG = '[EXAMPLE - EXTENDED SAMPLE]';
+
+const playerConfig: PlayerConfiguration = {
+  license: THEO_LICENSE,
+  chromeless: true,
+  hlsDateRange: true,
+  libraryLocation: 'theoplayer',
+  cast: {
+    chromecast: {
+      appID: 'CC1AD845',
+    },
+    strategy: 'auto',
+  },
+  ui: {
+    language: 'en',
+  },
+  mediaControl: {
+    mediaSessionEnabled: true,
+  },
+};
+
+export const SampleExtendedScreen = () => {
+  const [player, setPlayer] = useState<THEOplayer | undefined>(undefined);
+  const chromeless = playerConfig?.chromeless ?? false;
+  const onPlayerReady = (player: THEOplayer) => {
+    console.log(LOG_TAG, 'THEOplayer is ready:', player.version);
+
+    // store player reference
+    setPlayer(player);
+
+    // setup debug logs
+    player.addEventListener<PlayerEventType>(PlayerEventType.SOURCE_CHANGE, console.log);
+    player.addEventListener<PlayerEventType>(PlayerEventType.LOADED_DATA, console.log);
+    player.addEventListener<PlayerEventType>(PlayerEventType.LOADED_METADATA, console.log);
+    player.addEventListener<PlayerEventType>(PlayerEventType.READYSTATE_CHANGE, console.log);
+    player.addEventListener<PlayerEventType>(PlayerEventType.PLAY, console.log);
+    player.addEventListener<PlayerEventType>(PlayerEventType.PLAYING, console.log);
+    player.addEventListener<PlayerEventType>(PlayerEventType.PAUSE, console.log);
+    player.addEventListener<PlayerEventType>(PlayerEventType.SEEKING, console.log);
+    player.addEventListener<PlayerEventType>(PlayerEventType.SEEKED, console.log);
+    player.addEventListener<PlayerEventType>(PlayerEventType.ENDED, console.log);
+
+    // configure autoplay
+    player.autoplay = true;
+
+    // set initial source
+    player.source = SOURCES[0].source;
+
+    // configure backgroundAudio
+    player.backgroundAudioConfiguration = { enabled: true, shouldResumeAfterInterruption: true };
+
+    // configure PiP
+    player.pipConfiguration = { startsAutomatically: true };
+  };
+
+  return (
+    <SafeAreaView style={styles.PLAYER_CONTAINER_STYLE}>
+      <THEOplayerView config={playerConfig} onPlayerReady={onPlayerReady}>
+        {player !== undefined && chromeless && (
+          <UiContainer
+            theme={{ ...DEFAULT_THEOPLAYER_THEME }}
+            player={player}
+            behind={<CenteredDelayedActivityIndicator size={50} />}
+            top={
+              <ControlBar>
+                <MediaCacheMenuButton>
+                  <MediaCacheDownloadButton />
+                  <MediaCachingTaskListSubMenu />
+                </MediaCacheMenuButton>
+                {/*This is a custom menu for source selection.*/}
+                <SourceMenuButton />
+                {!Platform.isTV && (
+                  <>
+                    <AirplayButton />
+                    <ChromecastButton />
+                  </>
+                )}
+                <LanguageMenuButton />
+                <SettingsMenuButton>
+                  {/*Note: quality selection is not available on iOS */}
+                  <QualitySubMenu />
+                  <PlaybackRateSubMenu />
+                  <BackgroundAudioSubMenu />
+                  <PiPSubMenu />
+                </SettingsMenuButton>
+              </ControlBar>
+            }
+            center={<CenteredControlBar left={<SkipButton skip={-10} />} middle={<PlayButton />} right={<SkipButton skip={30} />} />}
+            bottom={
+              <>
+                <ControlBar style={{ justifyContent: 'flex-start' }}>
+                  <CastMessage />
+                </ControlBar>
+                {
+                  /*Note: RNSlider is not available on tvOS */
+                  !(Platform.isTV && Platform.OS === 'ios') && (
+                    <ControlBar>
+                      <SeekBar />
+                    </ControlBar>
+                  )
+                }
+                <ControlBar>
+                  <MuteButton />
+                  <TimeLabel showDuration={true} />
+                  <Spacer />
+                  <PipButton />
+                  <FullscreenButton />
+                </ControlBar>
+              </>
+            }
+          />
+        )}
+      </THEOplayerView>
+    </SafeAreaView>
+  );
+};
+
+const styles = StyleSheet.create({
+  PLAYER_CONTAINER_STYLE: {
+    flex: 1,
+    backgroundColor: 'transparent',
+  },
+});